+++ Mentoring project/Mapper/MappingProfile.cs
@@ -1,10 +1,4 @@
 ﻿using AutoMapper;
-using Mentoring_project.DTO;
-using Mentoring_project.Entities;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
 
 namespace Mentoring_project.Mapper
 {
@@ -12,7 +6,7 @@ namespace Mentoring_project.Mapper
     {
         public MappingProfile()
         {
-            CreateMap<UserDTO, User>();
+          //  CreateMap<UserDTO, User>();
         }
     }
 }
