+++ MentoringProject.Infrastructure.IoC/MentoringProject.Infrastructure.IoC.csproj
@@ -1,0 +1,16 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>net5.0</TargetFramework>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="5.0.1" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\Mentoring_project.Business\Mentoring_project.Business.csproj" />
+    <ProjectReference Include="..\Mentoring_project.Infrastructure.Data\Mentoring_project.Infrastructure.Data.csproj" />
+  </ItemGroup>
+
+</Project>
