+++ MentoringProject.Tests/UserServiceTest.cs
@@ -1,13 +1,8 @@
-using Mentoring_project.Entities;
-using Mentoring_project.Interfaces;
-using Mentoring_project.Repositories;
-using Mentoring_project.Services;
+using Mentoring_project.Business.Services;
+using Mentoring_project.Domain.Core.Entities;
+using Mentoring_project.Infrastructure.Data.Data;
 using MentoringProject.Tests;
-using Microsoft.EntityFrameworkCore;
-using Microsoft.Extensions.DependencyInjection;
-using Moq;
 using System;
-using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using Xunit;
