+++ Mentoring_project.Domain.Interfaces/Repositories/IRepository.cs
@@ -1,9 +1,10 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Text;
 using System.Threading.Tasks;
 
-namespace Mentoring_project.Repositories
+namespace Mentoring_project.Domain.Interfaces.Repositories
 {
     public interface IRepository<T> where T : class
     {
