+++ Mentoring_project.Domain/Entities/User.cs
@@ -1,9 +1,10 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Text;
 using System.Threading.Tasks;
 
-namespace Mentoring_project.Entities
+namespace Mentoring_project.Domain.Core.Entities
 {
     public class User
     {
