+++ Mentoring_project.Infrastructure.Data/Data/UnitOfWork.cs
@@ -1,23 +1,24 @@
-﻿using Mentoring_project.Entities;
-using Microsoft.EntityFrameworkCore;
+﻿using Mentoring_project.Domain.Core.Entities;
+using Mentoring_project.Domain.Interfaces.Repositories;
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Text;
 using System.Threading.Tasks;
 
-namespace Mentoring_project.Repositories
+namespace Mentoring_project.Infrastructure.Data.Data
 {
     public class UnitOfWork : IUnitOfWork
     {
         private readonly DbProjectContext _db;
-        private UserRepository userRepository;
+        private IRepository<User> userRepository;
 
         public UnitOfWork(DbProjectContext db)
         {
             _db = db;
         }
 
-        public UserRepository UserRepository
+        public IRepository<User> UserRepository
         {
             get
             {
@@ -37,4 +38,5 @@ namespace Mentoring_project.Repositories
             await _db.SaveChangesAsync();
         }
     }
+
 }
