+++ Mentoring_project.Infrastructure.Data/Data/UserRepository.cs
@@ -1,11 +1,13 @@
-﻿using Mentoring_project.Entities;
+﻿using Mentoring_project.Domain.Core.Entities;
+using Mentoring_project.Domain.Interfaces.Repositories;
 using Microsoft.EntityFrameworkCore;
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Text;
 using System.Threading.Tasks;
 
-namespace Mentoring_project.Repositories
+namespace Mentoring_project.Infrastructure.Data.Data
 {
     public class UserRepository : IRepository<User>
     {
